Make unknown space non-lethal
This change was made primarily to benefit the costmap accumulator.

The costmap accumulator makes a sort of "union" with a local costmap
and a larger blank map (matching our static map's size). We will
occasionally clear the local costmap to get rid of potentially stale
data. Currently, we replace stale data with a value that represents
"nothing".

This works well enough for navigation, but that data is
still potentially relevant for creating a large map over time.
If we use unknown values, we can simply not union unknown values.